Implements tests for Gaussian churn generator (#51)
* adds min and max range for gaussian generator as well as comprehensive doc

* adds tests for gaussian generator

* changes returned type of ints to double for churn

* wip sliding window test

* removes sliding window test

* fixes the tests

* fixes lint and updates java docs

* fixes tests

* increases tolerance threshold

+++ src/main/java/network/MiddleLayer.java
@@ -94,9 +94,10 @@ public class MiddleLayer {
     Integer port = fullAddress.getValue();
 
     // sleep for the simulated duration
-    int sleepTime = this.latencyGenerator.getSimulatedLatency(nodeId, destinationId, true);
+    double sleepTime = this.latencyGenerator.getSimulatedLatency(nodeId, destinationId, true);
     try {
-      Thread.sleep(sleepTime);
+      // TODO: fix this, we don't need to sleep!
+      Thread.sleep((int) sleepTime);
     } catch (InterruptedException ex) {
       this.logger.fatal("failed to sleep thread for the simulated delay, sleep time {}", sleepTime, ex);
     }
